@@ -1,55 +1,70 @@
 {
-  "name": "SystemChecker_Administration",
-  "version": "0.0.0",
+  "name": "SystemChecker.Administration",
+  "private": true,
+  "version": "1.0.0",
   "scripts": {
-    "test": "karma start ClientApp/test/karma.conf.js"
+    "test": "gulp test"
   },
   "dependencies": {
-    "@angular/animations": "4.1.2",
-    "@angular/common": "4.1.2",
-    "@angular/compiler": "4.1.2",
-    "@angular/core": "4.1.2",
-    "@angular/forms": "4.1.2",
-    "@angular/http": "4.1.2",
-    "@angular/platform-browser": "4.1.2",
-    "@angular/platform-browser-dynamic": "4.1.2",
-    "@angular/platform-server": "4.1.2",
-    "@angular/router": "4.1.2",
-    "@types/node": "7.0.18",
+    "@angular/animations": "4.2.3",
+    "@angular/common": "^4.2.3",
+    "@angular/compiler": "4.2.3",
+    "@angular/core": "4.2.3",
+    "@angular/forms": "4.2.3",
+    "@angular/http": "4.2.3",
+    "@angular/platform-browser": "4.2.3",
+    "@angular/platform-browser-dynamic": "4.2.3",
+    "@angular/platform-server": "4.2.3",
+    "@angular/router": "4.2.3",
+    "@types/core-js": "^0.9.41",
+    "@types/extract-text-webpack-plugin": "^2.1.0",
+    "@types/node": "8.0.0",
+    "@types/webpack": "^2.2.16",
+    "angular2-moment": "^1.4.0",
     "angular2-template-loader": "0.6.2",
-    "aspnet-prerendering": "^2.0.5",
     "aspnet-webpack": "^1.0.29",
     "awesome-typescript-loader": "3.1.3",
     "bootstrap": "3.3.7",
+    "core-js": "^2.4.1",
     "css": "2.2.1",
-    "css-loader": "0.28.1",
-    "es6-shim": "0.35.3",
+    "css-loader": "0.28.4",
+    "del": "^3.0.0",
     "event-source-polyfill": "0.0.9",
     "expose-loader": "0.7.3",
-    "extract-text-webpack-plugin": "2.1.0",
-    "file-loader": "0.11.1",
+    "extract-text-webpack-plugin": "^2.1.2",
+    "file-loader": "0.11.2",
+    "gulp": "^3.9.1",
+    "gulp-run": "^1.7.1",
+    "hammerjs": "^2.0.8",
     "html-loader": "0.4.5",
-    "isomorphic-fetch": "2.2.1",
+    "ignore-loader": "^0.1.2",
     "jquery": "3.2.1",
-    "json-loader": "0.5.4",
-    "preboot": "4.5.2",
-    "raw-loader": "0.5.1",
+    "moment": "^2.18.1",
+    "ng2-dragula": "^1.5.0",
+    "node-sass": "^4.5.3",
+    "pace-progress": "^1.0.2",
+    "primeng": "^4.1.0-rc.2",
     "reflect-metadata": "0.1.10",
-    "rxjs": "5.4.0",
-    "style-loader": "0.17.0",
+    "run-sequence": "^1.2.2",
+    "rxjs": "5.4.1",
+    "sass-loader": "^6.0.6",
+    "style-loader": "^0.18.2",
     "to-string-loader": "1.1.5",
-    "typescript": "2.3.2",
-    "url-loader": "0.5.8",
-    "webpack": "2.5.1",
+    "ts-node": "^3.0.6",
+    "tslint": "^5.4.3",
+    "tslint-language-service": "^0.9.6",
+    "typescript": "^2.3.4",
+    "url-loader": "0.5.9",
+    "webpack": "^2.6.1",
+    "webpack-bundle-analyzer": "^2.8.2",
     "webpack-hot-middleware": "2.18.0",
-    "webpack-merge": "4.1.0",
-    "zone.js": "0.8.10"
+    "zone.js": "^0.8.12"
   },
   "devDependencies": {
-    "@types/chai": "3.5.2",
-    "@types/jasmine": "2.5.47",
-    "chai": "3.5.0",
-    "jasmine-core": "2.6.1",
+    "@types/chai": "4.0.0",
+    "@types/jasmine": "2.5.52",
+    "chai": "4.0.2",
+    "jasmine-core": "2.6.4",
     "karma": "1.7.0",
     "karma-chai": "0.1.0",
     "karma-chrome-launcher": "2.1.1",
