@@ -1,8 +1,7 @@
-import { Component } from '@angular/core';
+import { Component } from "@angular/core";
 
 @Component({
-    selector: 'home',
-    templateUrl: './home.component.html'
+    selector: "home",
+    templateUrl: "./home.component.html",
 })
-export class HomeComponent {
-}
+export class HomeComponent { }
