@@ -1,16 +1,13 @@
-// Karma configuration file, see link for more information
-// https://karma-runner.github.io/0.13/config/configuration-file.html
-
 module.exports = function (config) {
     config.set({
         basePath: '.',
         frameworks: ['jasmine'],
         files: [
             '../../wwwroot/dist/vendor.js',
-            './boot-tests.ts'
+            '../../wwwroot/dist/boot-tests.js'
         ],
         preprocessors: {
-            './boot-tests.ts': ['webpack']
+            '*.ts': ['webpack']
         },
         reporters: ['progress'],
         port: 9876,
@@ -19,8 +16,8 @@ module.exports = function (config) {
         autoWatch: true,
         browsers: ['Chrome'],
         mime: { 'application/javascript': ['ts','tsx'] },
-        singleRun: false,
-        webpack: require('../../webpack.config.js')().filter(config => config.target !== 'node'), // Test against client bundle, because tests run in a browser
+        singleRun: true,
+        webpack: require('../../webpack.config.js'),
         webpackMiddleware: { stats: 'errors-only' }
     });
 };
