@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 
-namespace SystemChecker_Administration.Controllers
+namespace SystemChecker.Administration.Controllers
 {
     public class HomeController : Controller
     {
