@@ -1,10 +1,1 @@
-@{
-    ViewData["Title"] = "Home Page";
-}
-
-<app asp-prerender-module="ClientApp/dist/main-server">Loading...</app>
-
-<script src="~/dist/vendor.js" asp-append-version="true"></script>
-@section scripts {
-    <script src="~/dist/main-client.js" asp-append-version="true"></script>
-}
+<app>Loading..</app>
