@@ -1,15 +1,15 @@
-<!DOCTYPE html>
+﻿<!DOCTYPE html>
 <html>
     <head>
         <meta charset="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>@ViewData["Title"] - SystemChecker_Administration</title>
+        <title>SystemChecker.Administration</title>
         <base href="/" />
         <link rel="stylesheet" href="~/dist/vendor.css" asp-append-version="true" />
+        <script src="~/dist/vendor.js" asp-append-version="true" defer></script>
+        <script src="~/dist/main.js" asp-append-version="true" defer></script>
     </head>
     <body>
         @RenderBody()
-
-        @RenderSection("scripts", required: false)
     </body>
 </html>
