@@ -1,3 +1,3 @@
-@using SystemChecker_Administration
+@using SystemChecker.Administration
 @addTagHelper "*, Microsoft.AspNetCore.Mvc.TagHelpers"
 @addTagHelper "*, Microsoft.AspNetCore.SpaServices"
